+++ CHANGELOG.md
@@ -6,8 +6,18 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
 
+<<<<<<< HEAD
 ### Added
 - check-windows-directory.ps1, added plugin to check if a directory exist (@patricewhite).
+=======
+## [2.7.0] - 2018-05-09
+### Changed
+- check-windows-event-log.ps1, added plugin to check for pattern and returns the number criticals and warnings that match that pattern (@patricewhite)
+
+## [2.6.0] - 2018-05-09
+### Added
+- check-windows-log.ps1, added plugin to check for pattern in log file (@patricewhite).
+>>>>>>> bc6443ff2c6ef57497c88a1d664879cc1c587c06
 
 ### Changed
 - README.md, changed to include new plugin (@patricewhite).
@@ -156,7 +166,9 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.5.0...HEAD
+[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.7.0...HEAD
+[2.7.0]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.6.0...2.7.0
+[2.6.0]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.5.0...2.6.0
 [2.5.0]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.4.1...2.5.0
 [2.4.1]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.4.0...2.4.1
 [2.4.0]: https://github.com/sensu-plugins/sensu-plugins-Windows/compare/2.3.0...2.4.0